Small, important change to logic related to 'none' action and outputs.
Going back and forth on this, but I believe 'output' actions can be 'none' don't have to be upped to 'copy'. Files in the working directory are handled differently by the LWR and will are automatically copied over in that case. output_task files however are copied over by Galaxy, so these should probably remain 'copy' (*though this case remains untested*).