Parse destination_params for submit_params.
This feels some what wrong because a destination is describing both which LWR server and manager to use as well as the parameters to be used. This is a Galaxyism though - same holds true for all destinations the LWR justs feels heavier because we are constructing a client object per request.