Implement variant of client that talks directly to job manager instead of using HTTP server.
This could hopefully allow the LWR to be nested right inside of Galaxy someday, in the interim it does boost reported test coverage from 75% to 84% because of the way the test web server is started, none of the Python code was being instrumented properly. There is now a 'direct' communication variant of each of these tests resulting in the increased test coverage.