Enhance setup and finish message for greater Galaxy integration.
If GALAXY_HOME is set, let client know so that paths/scripts can be constructed against it. Pass list of working directory contents back to client at end of job - this is potentially useful as a step toward remote metadata setting and handling of tools with a variable number of outputs.