Allow launch to take a list of serialized tool requirements.
Build a Galaxy 'DependencyManager' in LwrApp and utilize it in managers to resolve these requirements.

A more correct approach would probably be to read the tool on LWR side, but this mode is potentially very useful for LWR instances not bound to toolboxes (i.e. every LWR instance I know about).