Adjustments for simplification of input and input_extra path types in previous commit.
Turns out legacy LWR servers are going to make a distinction between input and input_extra (which just eliminated). Hack for now is to upload everything as an 'extra input'. Really both LWR upload methods are deprecated to some degree in lieu of a generic upload method that should now work properly.