Eliminate references to lwr.util, it should all go through galaxy.util now.
Added following comment to galaxy/util/__init__.py

This file is a mess, it is a merge of random stuff that is in galaxy.util and
stuff that was in lwr.util. This should be reworked to only contain stuff in
galaxy.util and the rest should be moved into galaxy.util.lwr_io or something
like that.