WIP toward CLI baed submission of asynchronous jobs.
Idea is to allow one to submit jobs to remote hosts without formal LWR server running. This changeset starts work on an LWR client variant allowing this by bringing in the Galaxy's CLI runner to remotely execute shell commands. (Alternative remote execution mechanisms such as fabric or ansible may also be awesome to add in here.)

Much work still needs to be done in terms of creating the script referenced in this changeset to spin up the LWR, stage, submit or run directly a job, unstage, and connect to a MQ to register completion.

Once that is done - an alternative method of operation to target a new Galaxy API endpoint could be cool as well (so one could do this without the message queue in place).