Handle unlinking of SSH key instead of relying on tempfile
Perhaps it's a version/environment difference, but relying on tempfile to
unlink the file when done seems to not be working. I imagine it is because
tempfile unlinks the handle when I called f.close(), so we're just going to
handle unlinking ourselves and solve that problem permanently. This has the
added benefit of us knowing that the file will be removed as soon as the
transfer is complete. RSA keys on disk for a smaller amount of time "feels"
better.